add EventRow
